+++ CarRepairWorkshops.Application/Repairs/Commands/DeleteRepair/DeleteRepairCommandHandler.cs
@@ -2,6 +2,7 @@
 using CarRepairWorkshops.Application.Repairs.Commands.CreateRepair;
 using CarRepairWorkshops.Domain.Entities;
 using CarRepairWorkshops.Domain.Exceptions;
+using CarRepairWorkshops.Domain.Interfaces;
 using CarRepairWorkshops.Domain.Repositories;
 using MediatR;
 using Microsoft.Extensions.Logging;
@@ -14,7 +15,10 @@ using System.Threading.Tasks;
 namespace CarRepairWorkshops.Application.Repairs.Commands.DeleteRepair
 {
     internal class DeleteRepairCommandHandler(ILogger<CreateRepairCommandHandler> logger,
-    IMapper mapper, IRepairRepository repairRepository)
+    IMapper mapper, IRepairRepository repairRepository,
+    ICarRepairWorkshopsAuthorizationService authorizationService,
+    ICarRepairWorkshopsRepository workshopsRepository,
+    ICarsRepository carsRepository)
         : IRequestHandler<DeleteRepairCommand>
     {
         public async Task Handle(DeleteRepairCommand request, CancellationToken cancellationToken)
@@ -22,6 +26,12 @@ namespace CarRepairWorkshops.Application.Repairs.Commands.DeleteRepair
             logger.LogInformation("Deleting repair by given id: {repairId}", request.RepairId);
             var repair = await repairRepository.GetRepairByIdAsync(request.RepairId);
             if(repair == null) throw new NotFoundException(nameof(Repair), request.RepairId.ToString());
+
+            var car = await carsRepository.GetByIdAsync(repair.CarId);
+            var workshop = await workshopsRepository.GetById(car.CarRepairWorkshopId);
+
+            if (!authorizationService.AuthorizeMechanic(workshop)) throw new ForbidException();
+
 
             await repairRepository.DeleteEntity(repair);
         }
